SQL Hulk spidering tool and database creation script
The powershell script will spider a list of computers put in a servers.txt file (One server per line) and enumerate everything you could possibly want in your SQL environment.  You will need to create a database first on a server using the creation script, and make sure you change the values in the powershell script to reflect the server, as well as your corporate domain name so it can be crawled properly.